% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadData.R
\name{DownloadData}
\alias{DownloadData}
\title{Download additional GScluster Data}
\usage{
DownloadData('human')
}
\arguments{
\item{species}{text value for which species PPI data will downloaded.
possible value is : arabidopsis, celegans, eColi, fly, human, mouse, rat, yeast, zebrafish}
}
\description{
Download additional GScluster Data
}
