% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGSASeqData.R
\name{GetGSASeqData}
\alias{GetGSASeqData}
\title{Transform GSA-Seq result to use GScluster}
\usage{
GetGSASeqData(filename)
}
\arguments{
\item{filename}{directory of GSA-SNP2 result file}
}
\description{
Transform GSA-Seq result to use GScluster
}
