% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGSASNP2Data.R
\name{GetGSASNP2Data}
\alias{GetGSASNP2Data}
\title{Transform GSA-SNP2 result to use GScluster}
\usage{
GetGSASNP2Data(filename)
}
\arguments{
\item{filename}{directory of GSA-SNP2 result file}
}
\description{
Transform GSA-SNP2 result to use GScluster
}
