% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GScluster.R
\name{GScluster}
\alias{GScluster}
\title{Run GScluster}
\usage{
GScluster() # will run demo data
}
\arguments{
\item{GSAresult}{R object contain GSAresult , 3 (name,member,qvalue) or 4 (+direction) column matrix}

\item{GeneScores}{R object contain DE analysis result, 2 (genesymbol, score) column matrix}

\item{PPI}{Matrix with rowname, colname : genesymbol and value should 0 ~ 1. if null, STRING will be used.}

\item{Species}{A character for species, default is Human.
possible value is : (A)rabidopsis, (C).elegans, (E).Coli, (F)ly, (H)uman, (M)ouse, (R)at, (Y)east, (Z)ebrafish}

\item{alpha}{A numeric value between 0(No network) ~ 1 (Full network). weight for network in pMM Distance. default is 1}

\item{GsQCutoff}{A numeric value, threshold to use geneset. default is 0.25}

\item{GQCutoff}{A numeric value, threshold to use gene. default is 0.25}
}
\description{
Run GScluster
}
