% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGSASeqData.R
\name{GetGSASeqData}
\alias{GetGSASeqData}
\title{Transform GSA-Seq result to use GScluster}
\usage{
GetGSASeqData(GSA, Score)
}
\arguments{
\item{GSA}{directory of GSA-seq result file ( GSA result )}

\item{Score}{directory of GSA-seq result file ( Gene score )}
}
\description{
Transform GSA-Seq result to use GScluster
}
